refactored format-apa.ts to sperate files into more defined scopes
